requirements.txt
# Core dependencies
fastapi==0.104.1
aiogram==3.2.0
sqlalchemy==2.0.23
alembic==1.13.1

# Database drivers
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Rasa integration
rasa==3.6.13
rasa-sdk==3.6.2

# File processing
openpyxl==3.1.2
pandas==2.1.4
Pillow==10.1.0

# HTTP and utilities
httpx==0.25.2
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Web server
uvicorn[standard]==0.24.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
celery==5.3.4
redis==5.0.1
